examples/requirements.txt
datasets
fastcore
minai
torcheval